refactor: - add close button in login dialog - remove login function to toolbar component - update test
