+++ src/app/auth/components/login/login.component.ts
@@ -1,7 +1,5 @@
 import { Component } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
-import { AuthService } from '@auth/services/auth.service';
-import { SnackbarService } from '@lib/services/snackbar/snackbar.service';
 
 @Component({
   selector: 'dv-login',
@@ -16,15 +14,6 @@ export class LoginComponent {
   hidePassword = true;
 
   constructor(
-    private readonly fb: FormBuilder,
-    private readonly authService: AuthService,
-    private readonly snackbar: SnackbarService
+    private readonly fb: FormBuilder
   ) { }
-
-  onSubmit(): void {
-    this.authService.login(this.loginForm.value).subscribe({
-      next: () => this.snackbar.success('Login successfully!'),
-      error: () => this.snackbar.error('Something went wrong. Please try again!')
-    });
-  }
 }
